refactor: MobileNewsCard 분리
